Make the `ToTemplate` function shared
